feat: advanced fuzzing infrastructure improvements
## Phase 1: Instant Coverage Wins

- `cmake/sanitize.cmake`: add `trace-cmp` to `COVERAGE_FLAGS` and
  `WITHOUT_COVERAGE_FLAGS` so libFuzzer automatically solves magic-byte
  comparisons in `BinaryTypeIndex`, `CompressionMethodByte`, protocol
  version checks and similar switch statements with zero changes to any
  fuzz target
- All `tests/fuzz/*.options`: add `use_value_profile = 1` to every
  `[libfuzzer]` section to track input-dependent comparison values and
  help solve multi-byte checks

## Security fix: WASM bounds-check integer overflow

- `src/Interpreters/WebAssembly/WasmTimeRuntime.cpp`: fix integer
  overflow in `getMemory` where `ptr + size` (both `uint32_t`) could
  wrap around and bypass the bounds check, allowing host-memory access
  from WASM guest code.  Widen both operands to `uint64_t` before
  addition so `ptr=0xFFFFFFFF, size=2` correctly throws an exception.

## Phase 2: New fuzz targets

### Compression codecs (Gorilla, T64, FPC, Multiple)

Three codecs with complex bit-manipulation and the dangerous
`CompressionCodecMultiple` (attacker-controlled codec chaining via
`source[0]` = chain length, `source[1..N]` = `CompressionMethodByte`)
had no dedicated fuzzers.  Added:

- `src/Compression/fuzzers/gorilla_decompress_fuzzer.cpp`
- `src/Compression/fuzzers/t64_decompress_fuzzer.cpp`
- `src/Compression/fuzzers/fpc_decompress_fuzzer.cpp`
- `src/Compression/fuzzers/multiple_decompress_fuzzer.cpp`

Each follows the `delta_decompress_fuzzer` pattern; Gorilla/T64/FPC use
`CompressionCodecFactory::instance().get(method_byte)` since these
codecs read their own parameters from the compressed data header, making
the construction-time parameters irrelevant for decompression fuzzing.

### `decodeDataType` binary encoding fuzzer

`src/DataTypes/DataTypesBinaryEncoding.cpp` performs recursive type
construction from a binary stream with ~55 type indices.  Added:

- `src/DataTypes/fuzzers/decode_data_type_fuzzer.cpp`: feeds raw bytes
  to `decodeDataType(ReadBuffer&)` and, on success, exercises the
  decoded type's default serialization path
- `tests/fuzz/dictionaries/binary_types.dict`: all BinaryTypeIndex bytes
  0x00-0x36 plus common VarUInt encodings

### `SerializationLowCardinality` bulk deserialization fuzzer

`SerializationLowCardinality` has complex index arithmetic where a
dictionary index > dictionary size causes OOB access.  Added:

- `src/DataTypes/fuzzers/serialization_lowcardinality_fuzzer.cpp`:
  exercises `deserializeBinaryBulkStatePrefix` +
  `deserializeBinaryBulkWithMultipleStreams` for both
  `LowCardinality(String)` and `LowCardinality(Nullable(UInt64))`,
  toggling `native_format` via the auxiliary header byte

## Phase 3: Dictionaries and build

- `tests/fuzz/dictionaries/compression.dict`: `CompressionMethodByte`
  values, data-bytes-size hints, Gorilla/T64/FPC header patterns, and
  `CompressionCodecMultiple` chain examples
- `tests/fuzz/build.sh`: also copy `tests/fuzz/dictionaries/*.dict` to
  `$OUT/` so the new `.options` files can reference them by name

+++ src/Interpreters/WebAssembly/WasmTimeRuntime.cpp
@@ -166,7 +166,7 @@ public:
     uint8_t * getMemory(WasmPtr ptr, WasmSizeT size) override
     {
         auto memory_span = getMemory().data(store);
-        if (ptr + size >= memory_span.size())
+        if (static_cast<uint64_t>(ptr) + static_cast<uint64_t>(size) >= memory_span.size())
         {
             throw Exception(
                 ErrorCodes::WASM_ERROR,
